sycl: fix IGC compiler crash by using standard SYCL 2020 permute_group_by_xor
The dpct::permute_sub_group_by_xor function was using sycl::select_from_group
which requires the GroupNonUniformArithmetic SPIR-V capability (operand 6146).
This caused Intel Graphics Compiler (IGC) 2025.3.2 to crash with a floating
point exception during AOT compilation for DG2 GPUs.

Fix by adding a fast path that uses the standard SYCL 2020
sycl::permute_group_by_xor function when logical_sub_group_size >= physical
subgroup size (the common case). This avoids the problematic select_from_group
call while retaining the fallback for edge cases with smaller logical subgroups.

Amp-Thread-ID: https://ampcode.com/threads/T-019bf290-1c40-7507-a320-4f346f409873

+++ ggml/src/ggml-sycl/mmvq.cpp
@@ -49,12 +49,12 @@ static void mul_mat_vec_q_reorder(const void * __restrict__ vx, const void * __r
     }
 
     // sum up partial sums and write back result
-    bool has_work = (nd_item.get_local_id(2) / block_elements_per_subgroup) < blocks_per_row;
-    
+    bool has_work = static_cast<int>(nd_item.get_local_id(2) / block_elements_per_subgroup) < blocks_per_row;
+
 #pragma unroll
     for (int mask = QK_WARP_SIZE / 2; mask > 0; mask >>= 1) {
         float other_sum = dpct::permute_sub_group_by_xor(nd_item.get_sub_group(), partial_sum, mask);
-        bool other_has_work = ((nd_item.get_local_id(2) ^ mask) / block_elements_per_subgroup) < blocks_per_row;
+        bool other_has_work = static_cast<int>((nd_item.get_local_id(2) ^ mask) / block_elements_per_subgroup) < blocks_per_row;
         partial_sum = has_work && other_has_work ? partial_sum + other_sum : (has_work ? partial_sum : other_sum);
         has_work = has_work || other_has_work;
     }
@@ -107,12 +107,12 @@ static void mul_mat_vec_q_K_reorder(const void * __restrict__ vx, const void * _
     }
 
     // sum up partial sums and write back result
-    bool has_work = (nd_item.get_local_id(2) / block_elements_per_subgroup) < blocks_per_row;
-    
+    bool has_work = static_cast<int>(nd_item.get_local_id(2) / block_elements_per_subgroup) < blocks_per_row;
+
 #pragma unroll
     for (int mask = QK_WARP_SIZE / 2; mask > 0; mask >>= 1) {
         float other_sum = dpct::permute_sub_group_by_xor(nd_item.get_sub_group(), partial_sum, mask);
-        bool other_has_work = ((nd_item.get_local_id(2) ^ mask) / block_elements_per_subgroup) < blocks_per_row;
+        bool other_has_work = static_cast<int>((nd_item.get_local_id(2) ^ mask) / block_elements_per_subgroup) < blocks_per_row;
         partial_sum = has_work && other_has_work ? partial_sum + other_sum : (has_work ? partial_sum : other_sum);
         has_work = has_work || other_has_work;
     }
